Refactored models into its own module
