+++ backend/models/generators/version.py
@@ -1,4 +1,4 @@
-from backend.resources.version import VersionModel
+from backend.models.version import VersionModel
 
 VersionGenerator = [
     VersionModel(version = "0.0.1")
