+++ backend/models/version.py
@@ -1,0 +1,5 @@
+from backend.database import db
+
+class VersionModel(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    version = db.Column(db.String(10))
